requirements.txt
torch
transformers
Pillow
numpy
sentencepiece
